+++ geo/src/algorithm/mod.rs
@@ -38,6 +38,10 @@ pub use concave_hull::ConcaveHull;
 pub mod contains;
 pub use contains::Contains;
 
+/// Convert the type of a geometry’s coordinate value.
+pub mod convert;
+pub use convert::{Convert, TryConvert};
+
 /// Calculate the convex hull of a `Geometry`.
 pub mod convex_hull;
 pub use convex_hull::ConvexHull;