+++ geo/CHANGES.md
@@ -29,6 +29,8 @@
   - Consolidate traits `TryMapCoords` into `MapCoords` and `TryMapCoordsInplace` into `MapCoordsInPlace`
 * Implement `ChamberlainDuquetteArea` for all geo types.
   * <https://github.com/georust/geo/pull/833>
+* Add `{Convert,TryConvert}` traits for coordinate value type conversion.
+  * <https://github.com/georust/geo/pull/836>
 
 ## 0.20.1